Update schema.dtd
Lines 16-20 are still questionable.

Is line 16 an area name? It can be renamed to area_name;

I don't know what line 17 means;

Line 18 is the name of the owner? Is that just their first name, or also the last name? 

Line 19 can be mistaken for a type as in int/string/double. And I don't know what it represents.

Line 20 - do we expect to see data like "red", or data like "#ff0000"? 

Some of those problems can be solved by keeping the DTD the same, but attaching documentation which explains it.